Add primitive error handling
This completes part 1 of project 2:
https://github.com/pingcap/talent-plan/blob/master/rust/projects/project-2/project.md#part-1-error-handling

This is a primitive step towards having my own error handling for
future implementations of log-structured storage, which will require
IO operations that may fail. I'm pretty sure I'll end up iterating
over the current error implementation, and changing it, but for now
it allows everything to compile, so until I go on and find a reason to
change it, it's good enough for this part.